+++ packages/@webex/internal-plugin-metrics/test/unit/spec/call-diagnostic/call-diagnostic-metrics.ts
@@ -320,6 +320,32 @@ describe('internal-plugin-metrics', () => {
         });
       });
 
+      it('should build identifiers correctly with a meeting that has meetingInfo with siteName', () => {
+        const res = cd.getIdentifiers({
+          mediaConnections: [
+            {mediaAgentAlias: 'mediaAgentAlias', mediaAgentGroupId: 'mediaAgentGroupId'},
+          ],
+          webexConferenceIdStr: 'webexConferenceIdStr',
+          globalMeetingId: 'globalMeetingId',
+          meeting: {...fakeMeeting, meetingInfo: {...fakeMeeting.meetingInfo, confIdStr: 'webexConferenceIdStr1', meetingId: 'globalMeetingId1', siteName: 'siteName1'}},
+        });
+
+        assert.deepEqual(res, {
+          correlationId: 'correlationId',
+          webexConferenceIdStr: 'webexConferenceIdStr1',
+          globalMeetingId: 'globalMeetingId1',
+          deviceId: 'deviceUrl',
+          locusId: 'url',
+          locusStartTime: 'lastActive',
+          locusUrl: 'locus/url',
+          mediaAgentAlias: 'mediaAgentAlias',
+          mediaAgentGroupId: 'mediaAgentGroupId',
+          orgId: 'orgId',
+          userId: 'userId',
+          webexSiteName: 'siteName1',
+        });
+      });
+
       it('should build identifiers correctly given webexConferenceIdStr', () => {
         const res = cd.getIdentifiers({
           correlationId: 'correlationId',
@@ -440,6 +466,7 @@ describe('internal-plugin-metrics', () => {
         const submitToCallDiagnosticsSpy = sinon.spy(cd, 'submitToCallDiagnostics');
         const generateClientEventErrorPayloadSpy = sinon.spy(cd, 'generateClientEventErrorPayload');
         const getIdentifiersSpy = sinon.spy(cd, 'getIdentifiers');
+        const getSubServiceTypeSpy = sinon.spy(cd, 'getSubServiceType');
         sinon.stub(cd, 'getOrigin').returns({origin: 'fake-origin'});
         const validatorSpy = sinon.spy(cd, 'validator');
         const options = {
@@ -481,6 +508,7 @@ describe('internal-plugin-metrics', () => {
             name: 'client.alert.displayed',
             userType: 'host',
             isConvergedArchitectureEnabled: undefined,
+            webexSubServiceType: undefined,
           },
           options
         );
@@ -505,6 +533,7 @@ describe('internal-plugin-metrics', () => {
             name: 'client.alert.displayed',
             userType: 'host',
             isConvergedArchitectureEnabled: undefined,
+            webexSubServiceType: undefined,
           },
           eventId: 'my-fake-id',
           origin: {
@@ -538,6 +567,7 @@ describe('internal-plugin-metrics', () => {
             name: 'client.alert.displayed',
             userType: 'host',
             isConvergedArchitectureEnabled: undefined,
+            webexSubServiceType: undefined,
           },
           eventId: 'my-fake-id',
           origin: {
@@ -556,22 +586,7 @@ describe('internal-plugin-metrics', () => {
           'call-diagnostic-events -> ',
           'CallDiagnosticMetrics: @submitClientEvent. Submit Client Event CA event.',
           `name: client.alert.displayed`,
-          `payload: undefined`,
-          `options: ${JSON.stringify(options)}`,
         ]);
-
-        assert.deepEqual(webexLoggerLogCalls[1].args, [
-          'call-diagnostic-events -> ',
-          'CallDiagnosticMetrics: @createClientEventObjectInMeeting. Creating in meeting event object.',
-          `name: client.alert.displayed`
-        ]);
-
-        assert.deepEqual(webexLoggerLogCalls[2].args, [
-          'call-diagnostic-events -> ',
-          'CallDiagnosticMetrics: @submitToCallDiagnostics. Preparing to send the request',
-          `finalEvent: {"eventPayload":{"eventId":"my-fake-id","version":1,"origin":{"origin":"fake-origin"},"originTime":{"triggered":"${now.toISOString()}","sent":"not_defined_yet"},"senderCountryCode":"UK","event":{"name":"client.alert.displayed","canProceed":true,"identifiers":{"correlationId":"correlationId","userId":"userId","deviceId":"deviceUrl","orgId":"orgId","locusUrl":"locus/url","locusId":"url","locusStartTime":"lastActive","mediaAgentAlias":"alias","mediaAgentGroupId":"1"},"eventData":{"webClientDomain":"whatever"},"userType":"host","loginType":"login-ci"}},"type":["diagnostic-event"]}`,
-        ]);
-
       });
 
       it('should submit client event successfully with correlationId, webexConferenceIdStr and globalMeetingId', () => {
@@ -657,22 +672,7 @@ describe('internal-plugin-metrics', () => {
           'call-diagnostic-events -> ',
           'CallDiagnosticMetrics: @submitClientEvent. Submit Client Event CA event.',
           `name: client.alert.displayed`,
-          `payload: undefined`,
-          `options: ${JSON.stringify(options)}`,
         ]);
-
-        assert.deepEqual(webexLoggerLogCalls[1].args, [
-          'call-diagnostic-events -> ',
-          'CallDiagnosticMetrics: @createClientEventObjectPreMeeting. Creating pre meeting event object.',
-          `name: client.alert.displayed`
-        ]);
-
-        assert.deepEqual(webexLoggerLogCalls[2].args, [
-          'call-diagnostic-events -> ',
-          'CallDiagnosticMetrics: @submitToCallDiagnostics. Preparing to send the request',
-          `finalEvent: {"eventPayload":{"eventId":"my-fake-id","version":1,"origin":{"origin":"fake-origin"},"originTime":{"triggered":"${now.toISOString()}","sent":"not_defined_yet"},"senderCountryCode":"UK","event":{"name":"client.alert.displayed","canProceed":true,"identifiers":{"correlationId":"correlationId","userId":"userId","deviceId":"deviceUrl","orgId":"orgId","locusUrl":"locus-url","webexConferenceIdStr":"webexConferenceIdStr1","globalMeetingId":"globalMeetingId1"},"eventData":{"webClientDomain":"whatever"},"loginType":"login-ci"}},"type":["diagnostic-event"]}`,
-        ]);
-
 
       });
 
@@ -812,6 +812,7 @@ describe('internal-plugin-metrics', () => {
             name: 'client.alert.displayed',
             userType: 'host',
             isConvergedArchitectureEnabled: undefined,
+            webexSubServiceType: undefined,
           },
           eventId: 'my-fake-id',
           origin: {
@@ -830,35 +831,14 @@ describe('internal-plugin-metrics', () => {
           'call-diagnostic-events -> ',
           'CallDiagnosticMetrics: @submitClientEvent. Submit Client Event CA event.',
           `name: client.alert.displayed`,
-          `payload: undefined`,
-          `options: ${JSON.stringify(options)}`,
-        ]);
 
+        ]);
 
         assert.deepEqual(webexLoggerLogCalls[1].args, [
-          'call-diagnostic-events -> ',
-          'CallDiagnosticMetrics: @prepareClientEvent. Error detected, attempting to map and attach it to the event...',
-          `name: client.alert.displayed`,
-          `rawError: ${options.rawError}`
-        ]);
-
-        assert.deepEqual(webexLoggerLogCalls[2].args, [
           'call-diagnostic-events -> ',
           'CallDiagnosticMetrics: @prepareClientEvent. Generated errors:',
           `generatedError: {"fatal":true,"shownToUser":false,"name":"other","category":"expected","errorCode":4029,"serviceErrorCode":2409005,"errorDescription":"StartRecordingFailed"}`
         ])
-
-        assert.deepEqual(webexLoggerLogCalls[3].args, [
-          'call-diagnostic-events -> ',
-          'CallDiagnosticMetrics: @createClientEventObjectInMeeting. Creating in meeting event object.',
-          `name: client.alert.displayed`
-        ]);
-
-        assert.deepEqual(webexLoggerLogCalls[4].args, [
-          'call-diagnostic-events -> ',
-          'CallDiagnosticMetrics: @submitToCallDiagnostics. Preparing to send the request',
-          `finalEvent: {"eventPayload":{"eventId":"my-fake-id","version":1,"origin":{"origin":"fake-origin"},"originTime":{"triggered":"${now.toISOString()}","sent":"not_defined_yet"},"senderCountryCode":"UK","event":{"name":"client.alert.displayed","canProceed":true,"identifiers":{"correlationId":"correlationId","userId":"userId","deviceId":"deviceUrl","orgId":"orgId","locusUrl":"locus/url","locusId":"url","locusStartTime":"lastActive","mediaAgentAlias":"alias","mediaAgentGroupId":"1","webexConferenceIdStr":"webexConferenceIdStr1","globalMeetingId":"globalMeetingId1"},"errors":[{"fatal":true,"shownToUser":false,"name":"other","category":"expected","errorCode":4029,"serviceErrorCode":2409005,"errorDescription":"StartRecordingFailed"}],"eventData":{"webClientDomain":"whatever"},"userType":"host","loginType":"login-ci"}},"type":["diagnostic-event"]}`,
-        ]);
       });
 
       it('it send the raw error message if meetingId provided', () => {
@@ -911,6 +891,7 @@ describe('internal-plugin-metrics', () => {
             name: 'client.alert.displayed',
             userType: 'host',
             isConvergedArchitectureEnabled: undefined,
+            webexSubServiceType: undefined,
           },
           eventId: 'my-fake-id',
           origin: {
@@ -929,37 +910,12 @@ describe('internal-plugin-metrics', () => {
           'call-diagnostic-events -> ',
           'CallDiagnosticMetrics: @submitClientEvent. Submit Client Event CA event.',
           `name: client.alert.displayed`,
-          `payload: undefined`,
-          `options: ${JSON.stringify(options)}`,
         ]);
 
         assert.deepEqual(webexLoggerLogCalls[1].args, [
           'call-diagnostic-events -> ',
-          'CallDiagnosticMetrics: @prepareClientEvent. Error detected, attempting to map and attach it to the event...',
-          `name: client.alert.displayed`,
-          `rawError: ${JSON.stringify({
-            message: options.rawError.message,
-            name: options.rawError.name,
-            stack: options.rawError.stack,
-          })}`,
-        ]);
-
-        assert.deepEqual(webexLoggerLogCalls[2].args, [
-          'call-diagnostic-events -> ',
           'CallDiagnosticMetrics: @prepareClientEvent. Generated errors:',
           `generatedError: {"fatal":true,"shownToUser":false,"name":"other","category":"other","errorCode":9999,"serviceErrorCode":9999,"errorDescription":"UnknownError"}`,
-        ]);
-
-        assert.deepEqual(webexLoggerLogCalls[3].args, [
-          'call-diagnostic-events -> ',
-          'CallDiagnosticMetrics: @createClientEventObjectInMeeting. Creating in meeting event object.',
-          `name: client.alert.displayed`,
-        ]);
-
-        assert.deepEqual(webexLoggerLogCalls[4].args, [
-          'call-diagnostic-events -> ',
-          'CallDiagnosticMetrics: @submitToCallDiagnostics. Preparing to send the request',
-          `finalEvent: {"eventPayload":{"eventId":"my-fake-id","version":1,"origin":{"origin":"fake-origin"},"originTime":{"triggered":"${now.toISOString()}","sent":"not_defined_yet"},"senderCountryCode":"UK","event":{"name":"client.alert.displayed","canProceed":true,"identifiers":{"correlationId":"correlationId","userId":"userId","deviceId":"deviceUrl","orgId":"orgId","locusUrl":"locus/url","locusId":"url","locusStartTime":"lastActive","mediaAgentAlias":"alias","mediaAgentGroupId":"1"},"errors":[{"fatal":true,"shownToUser":false,"name":"other","category":"other","errorCode":9999,"serviceErrorCode":9999,"errorDescription":"UnknownError"}],"eventData":{"webClientDomain":"whatever","rawErrorMessage":"bad times"},"userType":"host","loginType":"login-ci"}},"type":["diagnostic-event"]}`,
         ]);
       });
 
@@ -1023,39 +979,15 @@ describe('internal-plugin-metrics', () => {
           'call-diagnostic-events -> ',
           'CallDiagnosticMetrics: @submitClientEvent. Submit Client Event CA event.',
           `name: client.alert.displayed`,
-          `payload: undefined`,
-          `options: ${JSON.stringify(options)}`,
         ]);
 
 
         assert.deepEqual(webexLoggerLogCalls[1].args, [
           'call-diagnostic-events -> ',
-          'CallDiagnosticMetrics: @prepareClientEvent. Error detected, attempting to map and attach it to the event...',
-          `name: client.alert.displayed`,
-          `rawError: ${JSON.stringify({
-            message: options.rawError.message,
-            name: options.rawError.name,
-            stack: options.rawError.stack,
-          })}`,
-        ]);
-
-        assert.deepEqual(webexLoggerLogCalls[2].args, [
-          'call-diagnostic-events -> ',
           'CallDiagnosticMetrics: @prepareClientEvent. Generated errors:',
           `generatedError: {"fatal":true,"shownToUser":false,"name":"other","category":"other","errorCode":9999,"serviceErrorCode":9999,"errorDescription":"UnknownError"}`,
         ]);
 
-        assert.deepEqual(webexLoggerLogCalls[3].args, [
-          'call-diagnostic-events -> ',
-          'CallDiagnosticMetrics: @createClientEventObjectPreMeeting. Creating pre meeting event object.',
-          `name: client.alert.displayed`
-        ]);
-
-        assert.deepEqual(webexLoggerLogCalls[4].args, [
-          'call-diagnostic-events -> ',
-          'CallDiagnosticMetrics: @submitToCallDiagnostics. Preparing to send the request',
-          `finalEvent: {"eventPayload":{"eventId":"my-fake-id","version":1,"origin":{"origin":"fake-origin"},"originTime":{"triggered":"${now.toISOString()}","sent":"not_defined_yet"},"senderCountryCode":"UK","event":{"name":"client.alert.displayed","errors":[{"fatal":true,"shownToUser":false,"name":"other","category":"other","errorCode":9999,"serviceErrorCode":9999,"errorDescription":"UnknownError"}],"canProceed":true,"identifiers":{"correlationId":"correlationId","userId":"userId","deviceId":"deviceUrl","orgId":"orgId","locusUrl":"locus-url"},"eventData":{"webClientDomain":"whatever","rawErrorMessage":"bad times"},"loginType":"login-ci"}},"type":["diagnostic-event"]}`,
-        ]);
       });
 
       it('it should send the raw error message if provided with correlationId', () => {
@@ -1121,34 +1053,14 @@ describe('internal-plugin-metrics', () => {
           'call-diagnostic-events -> ',
           'CallDiagnosticMetrics: @submitClientEvent. Submit Client Event CA event.',
           `name: client.alert.displayed`,
-          `payload: undefined`,
-          `options: ${JSON.stringify(options)}`,
         ]);
 
         assert.deepEqual(webexLoggerLogCalls[1].args, [
-          'call-diagnostic-events -> ',
-          'CallDiagnosticMetrics: @prepareClientEvent. Error detected, attempting to map and attach it to the event...',
-          `name: client.alert.displayed`,
-          `rawError: ${options.rawError}`,
-        ]);
-
-        assert.deepEqual(webexLoggerLogCalls[2].args, [
           'call-diagnostic-events -> ',
           'CallDiagnosticMetrics: @prepareClientEvent. Generated errors:',
           `generatedError: {"fatal":true,"shownToUser":false,"name":"other","category":"expected","errorCode":4029,"serviceErrorCode":2409005,"errorDescription":"StartRecordingFailed"}`,
         ]);
 
-        assert.deepEqual(webexLoggerLogCalls[3].args, [
-          'call-diagnostic-events -> ',
-          'CallDiagnosticMetrics: @createClientEventObjectPreMeeting. Creating pre meeting event object.',
-          `name: client.alert.displayed`,
-        ]);
-
-        assert.deepEqual(webexLoggerLogCalls[4].args, [
-          'call-diagnostic-events -> ',
-          'CallDiagnosticMetrics: @submitToCallDiagnostics. Preparing to send the request',
-          `finalEvent: {"eventPayload":{"eventId":"my-fake-id","version":1,"origin":{"origin":"fake-origin"},"originTime":{"triggered":"${now.toISOString()}","sent":"not_defined_yet"},"senderCountryCode":"UK","event":{"name":"client.alert.displayed","errors":[{"fatal":true,"shownToUser":false,"name":"other","category":"expected","errorCode":4029,"serviceErrorCode":2409005,"errorDescription":"StartRecordingFailed"}],"canProceed":true,"identifiers":{"correlationId":"correlationId","userId":"userId","deviceId":"deviceUrl","orgId":"orgId","locusUrl":"locus-url"},"eventData":{"webClientDomain":"whatever"},"loginType":"login-ci"}},"type":["diagnostic-event"]}`,
-        ]);
       });
 
       it('should include errors in payload if provided via payload', () => {
@@ -1204,6 +1116,7 @@ describe('internal-plugin-metrics', () => {
             name: 'client.alert.displayed',
             userType: 'host',
             isConvergedArchitectureEnabled: undefined,
+            webexSubServiceType: undefined,
           },
           eventId: 'my-fake-id',
           origin: {
@@ -1665,7 +1578,7 @@ describe('internal-plugin-metrics', () => {
           body: {},
           options: {headers: {}, url: 'https://example.com'},
         });
-        
+
         error.payloadOverrides = {
           shownToUser: true,
           category: 'expected'
@@ -1749,6 +1662,41 @@ describe('internal-plugin-metrics', () => {
       });
     });
 
+    describe('#getSubServiceType', () => {
+      it('returns subServicetype as PMR when PMR meeting', () => {
+       fakeMeeting.meetingInfo ={
+          webexScheduled: false,
+          pmr: true,
+          enableEvent: false,
+        }
+        assert.deepEqual(cd.getSubServiceType(fakeMeeting), 'PMR');
+      });
+
+      it('returns subServicetype as ScheduledMeeting when regular meeting', () => {
+        fakeMeeting.meetingInfo ={
+           webexScheduled: true,
+           pmr: false,
+           enableEvent: false,
+         }
+         assert.deepEqual(cd.getSubServiceType(fakeMeeting), 'ScheduledMeeting');
+       });
+
+       it('returns subServicetype as Webinar when meeting is Webinar', () => {
+        fakeMeeting.meetingInfo ={
+           webexScheduled: true,
+           pmr: false,
+           enableEvent: true,
+         }
+         assert.deepEqual(cd.getSubServiceType(fakeMeeting), 'Webinar');
+       });
+
+       it('returns subServicetype as undefined when correct parameters are not found', () => {
+        fakeMeeting.meetingInfo ={}
+         assert.deepEqual(cd.getSubServiceType(fakeMeeting), undefined);
+       });
+
+    });
+
     describe('#getIsConvergedArchitectureEnabled', () => {
       it('returns true if converged architecture is enabled', () => {
         fakeMeeting.meetingInfo = {enableConvergedArchitecture: true};
@@ -1805,6 +1753,7 @@ describe('internal-plugin-metrics', () => {
                       trigger: 'user-interaction',
                       userType: 'host',
                       isConvergedArchitectureEnabled: undefined,
+                      webexSubServiceType: undefined,
                     },
                     eventId: 'my-fake-id',
                     origin: {
@@ -1867,14 +1816,6 @@ describe('internal-plugin-metrics', () => {
             'call-diagnostic-events -> ',
             'CallDiagnosticMetrics: @buildClientEventFetchRequestOptions. Building request options object for fetch()...',
             `name: client.exit.app`,
-            `payload: {"trigger":"user-interaction","canProceed":false}`,
-            `options: ${JSON.stringify(options)}`,
-          ]);
-
-          assert.deepEqual(webexLoggerLogCalls[1].args, [
-            'call-diagnostic-events -> ',
-            'CallDiagnosticMetrics: @createClientEventObjectInMeeting. Creating in meeting event object.',
-            `name: client.exit.app`
           ]);
         });
       });